iot-manager/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
arrow==1.3.0
async-timeout==5.0.1
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
codemcp @ git+https://github.com/ezyang/codemcp@25a724ddba7091a669cc39e2f0540a542eb21470
EditorConfig==0.17.0
execnet==2.1.1
Flask==3.1.0
Flask-Cors==3.0.10
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mcp==1.6.0
mdurl==0.1.2
numpy==2.2.4
opencv-python==4.11.0.86
packaging==24.2
pandas==2.2.3
pluggy==1.5.0
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
Pygments==2.19.1
pytest==8.3.5
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
Redis-Sentinel-Url==1.0.1
rich==14.0.0
rq==2.3.1
rq-dashboard==0.8.2.2
ruff==0.11.5
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sse-starlette==2.2.1
starlette==0.46.1
toml==0.10.2
tomli==2.2.1
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
uvicorn==0.34.0
Werkzeug==3.1.3
